+++ CHANGELOG.md
@@ -32,6 +32,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - Labels Component
 - Breadcrumb Component
 - Text Input Field
+- Form Label Component
 
 ### Changed
 
@@ -72,3 +73,9 @@ Initial release of the Design System Package
 ### Added
 
 - Updated tailwind to version 3
+
+## [v1.20.0] - 2022-01-20
+
+### Added
+
+- Updated code base with "ds-" prefix